api/requirements.txt
anyio==3.6.2
cffi==1.15.1
click==8.1.3
cryptography==39.0.0
dnspython==2.3.0
ecdsa==0.18.0
email-validator==1.3.1
fastapi==0.89.1
h11==0.14.0
httptools==0.5.0
idna==3.4
passlib==1.7.4
pyasn1==0.4.8
pycparser==2.21
pydantic==1.10.4
pymongo==4.3.3
python-dotenv==0.21.0
python-jose==3.3.0
python-multipart==0.0.5
pytz==2022.7.1
PyYAML==6.0
rsa==4.9
six==1.16.0
sniffio==1.3.0
starlette==0.22.0
typing_extensions==4.4.0
uvicorn==0.20.0
uvloop==0.17.0
watchfiles==0.18.1
websockets==10.4
